refactor(annotations): move tag_attribution into the annotations app
The eval-tag attribution helper resolves who/what applied a tag for the generic
tag tooltip, which is an annotations concern. Moving it from evaluations to
annotations lets callers import the service from annotations and stop reaching
into the evaluations app. tag_views.py (which already imports evaluations.models)
now imports it at top level; remaining callers keep the local import to avoid
circular imports through evaluations.models.
