+++ apps/participants/views.py
@@ -51,7 +51,7 @@ def single_participant_home_context(team, context: dict, participant_id: int, ex
             .filter(participant=participant)
             .prefetch_related(chat_tagged_items_prefetch())
         )
-        from apps.evaluations.tag_attribution import attach_tag_attributions  # noqa: PLC0415 — avoid circular import
+        from apps.annotations.tag_attribution import attach_tag_attributions  # noqa: PLC0415 — avoid circular import
 
         attach_tag_attributions([session.chat for session in sessions])
         context["session_table"] = ChatbotSessionsTable(
